Testing controller in open loop.
Matlab Simulink includes all the files within the blackbox config file by
default, causing a horrible problem when you try to include a library. I
have changed the ieee_proposed library to be at work library.

Simulink is now in open loop. And I need to test better the controller
because it seems to be crap.